build(squirrel): generate delta release
Prior to this commit, the `--no-delta` flag was passed to the
`Squirrel.com --releasify ...` build command to skip building a delta
release in an effort to save time.  From empirical testing, it looks like
generating the delta release takes about 4 extra minutes (i.e., ~5.5
minutes instead of ~1.5 minutes without the delta release).  However, the
delta release may _dramatically_ reduce the download size required for
updating.  For example, a change to two plugin files resulted in a delta
release of ~6.6 MB as opposed to the full release of ~254 MB.

In this commit, remove the `--no-delta` flag from the Squirrel build
command to _include_ creation of delta releases.